+++ fastly/block_fastly_service_logging_https.go
@@ -34,10 +34,26 @@ func (h *HTTPSLoggingServiceAttributeHandler) Key() string {
 // GetSchema returns the resource schema.
 func (h *HTTPSLoggingServiceAttributeHandler) GetSchema() *schema.Schema {
 	blockAttributes := map[string]*schema.Schema{
+		"compression_codec": {
+			Type:             schema.TypeString,
+			Optional:         true,
+			Description:      `The codec used for compression of your logs. Valid values are zstd, snappy, and gzip. If the specified codec is "gzip", gzip_level will default to 3. To specify a different level, leave compression_codec blank and explicitly set the level using gzip_level. Specifying both compression_codec and gzip_level in the same API request will result in an error.`,
+			ValidateDiagFunc: validateLoggingCompressionCodec(),
+		},
 		"content_type": {
 			Type:        schema.TypeString,
 			Optional:    true,
 			Description: "Value of the `Content-Type` header sent with the request",
+		},
+		"gzip_level": {
+			Type:     schema.TypeInt,
+			Optional: true,
+			// NOTE: The default represents an unset value
+			// We use this instead of zero because the zero value for an int type is
+			// actually a valid value for the API. The API will attempt to default to
+			// zero if nothing is set by the user in their TF configuration.
+			Default:     -1,
+			Description: GzipLevelDescription,
 		},
 		"header_name": {
 			Type:        schema.TypeString,
@@ -188,7 +204,7 @@ func (h *HTTPSLoggingServiceAttributeHandler) Read(ctx context.Context, d *schem
 			return fmt.Errorf("error looking up HTTPS logging endpoints for (%s), version (%v): %s", d.Id(), serviceVersion, err)
 		}
 
-		hll := flattenHTTPS(remoteState)
+		hll := flattenHTTPS(remoteState, localState)
 
 		for _, element := range hll {
 			h.pruneVCLLoggingAttributes(element)
@@ -226,6 +242,12 @@ func (h *HTTPSLoggingServiceAttributeHandler) Update(ctx context.Context, d *sch
 	}
 	if v, ok := modified["content_type"]; ok {
 		opts.ContentType = gofastly.ToPointer(v.(string))
+	}
+	if v, ok := modified["compression_codec"]; ok {
+		opts.CompressionCodec = gofastly.ToPointer(v.(string))
+	}
+	if v, ok := modified["gzip_level"]; ok {
+		opts.GzipLevel = gofastly.ToPointer(v.(int))
 	}
 	if v, ok := modified["header_name"]; ok {
 		opts.HeaderName = gofastly.ToPointer(v.(string))
@@ -292,9 +314,32 @@ func (h *HTTPSLoggingServiceAttributeHandler) Delete(ctx context.Context, d *sch
 }
 
 // flattenHTTPS models data into format suitable for saving to Terraform state.
-func flattenHTTPS(remoteState []*gofastly.HTTPS) []map[string]any {
+func flattenHTTPS(remoteState []*gofastly.HTTPS, localState []any) []map[string]any {
 	var result []map[string]any
 	for _, resource := range remoteState {
+		// Avoid setting gzip_level to the API default of zero if originally unset.
+		// This avoids an unnecessary diff where the local state would have been
+		// updated to zero and so would be different from the -1 default set.
+		// As the user never set the attribute we don't want to show a diff to say
+		// it should be zero according to the API.
+		//
+		// NOTE: Ideally the local state would be updated when .Create() is called.
+		// e.g. we'd check if the value is -1 for gzip_level and set it in state as
+		// zero instead. This way we could avoid having to do this check here.
+		// The reason that's not possible (or not ideal at least) is because Create
+		// is called multiple times (once for each block defined in configuration)
+		// while the setting of the state must be done holistically, and so what
+		// that means is, if we did the above suggestion we would be resetting the
+		// entire state object multiple times, where as here we're only ever setting
+		// it once.
+		for _, s := range localState {
+			v := s.(map[string]any)
+			if resource.Name != nil && v["name"].(string) == *resource.Name && v["gzip_level"].(int) == -1 {
+				resource.GzipLevel = gofastly.ToPointer(v["gzip_level"].(int))
+				break
+			}
+		}
+
 		data := map[string]any{}
 
 		if resource.Name != nil {
@@ -315,11 +360,17 @@ func flattenHTTPS(remoteState []*gofastly.HTTPS) []map[string]any {
 		if resource.RequestMaxBytes != nil {
 			data["request_max_bytes"] = *resource.RequestMaxBytes
 		}
+		if resource.CompressionCodec != nil {
+			data["compression_codec"] = *resource.CompressionCodec
+		}
 		if resource.ContentType != nil {
 			data["content_type"] = *resource.ContentType
 		}
 		if resource.HeaderName != nil {
 			data["header_name"] = *resource.HeaderName
+		}
+		if resource.GzipLevel != nil {
+			data["gzip_level"] = *resource.GzipLevel
 		}
 		if resource.HeaderValue != nil {
 			data["header_value"] = *resource.HeaderValue
@@ -373,6 +424,7 @@ func (h *HTTPSLoggingServiceAttributeHandler) buildCreate(httpsMap any, serviceI
 
 	vla := h.getVCLLoggingAttributes(resource)
 	opts := gofastly.CreateHTTPSInput{
+		CompressionCodec:  gofastly.ToPointer(resource["compression_codec"].(string)),
 		ContentType:       gofastly.ToPointer(resource["content_type"].(string)),
 		Format:            gofastly.ToPointer(vla.format),
 		FormatVersion:     vla.formatVersion,
@@ -392,6 +444,14 @@ func (h *HTTPSLoggingServiceAttributeHandler) buildCreate(httpsMap any, serviceI
 		TLSHostname:       gofastly.ToPointer(resource["tls_hostname"].(string)),
 		URL:               gofastly.ToPointer(resource["url"].(string)),
 		ProcessingRegion:  gofastly.ToPointer(resource["processing_region"].(string)),
+	}
+
+	// NOTE: go-fastly v7+ expects a pointer, so TF can't set the zero type value.
+	// If we set a default value for an attribute, then it will be sent to the API.
+	// In some scenarios this can cause the API to reject the request.
+	// For example, configuring compression_codec + gzip_level is invalid.
+	if gl, ok := resource["gzip_level"].(int); ok && gl != -1 {
+		opts.GzipLevel = gofastly.ToPointer(gl)
 	}
 
 	// WARNING: The following fields shouldn't have an empty string passed.